renamed fake localizer to ground truth localizer
