+++ astrobee/launch/astrobee.launch
@@ -29,7 +29,7 @@
 
   <!-- Remaining options -->
   <arg name="output" default="log" />         <!-- Output to screen or log   -->
-  <arg name="fake_localizer" default="false" /> <!-- Use Fake Localizer      -->
+  <arg name="ground_truth_localizer" default="false" /> <!-- Use Ground Truth Localizer      -->
   <arg name="drivers" default="true" />       <!-- Should we launch drivers? -->
   <arg name="sim" default="local" />          <!-- SIM IP address            -->
   <arg name="llp" default="local" />          <!-- LLP IP address            -->
@@ -114,7 +114,7 @@
         <arg name="extra" value="$(arg extra)" />      <!-- Inject extra nodes   -->
         <arg name="debug" value="$(arg debug)" />      <!-- Debug node group     -->
         <arg name="output" value="$(arg output)" />
-        <arg name="fake_localizer" value="$(arg fake_localizer)" />
+        <arg name="ground_truth_localizer" value="$(arg ground_truth_localizer)" />
       </include>
 
       <!-- If we have specified a bag, then play it back into flight software-->
@@ -154,7 +154,7 @@
         <arg name="extra" value="$(arg extra)" />      <!-- Inject extra nodes   -->
         <arg name="debug" value="$(arg debug)" />      <!-- Debug node group     -->
         <arg name="output" value="$(arg output)" />
-        <arg name="fake_localizer" value="$(arg fake_localizer)" />
+        <arg name="ground_truth_localizer" value="$(arg ground_truth_localizer)" />
       </include>
 
       <!-- If we have specified a bag, then play it back into flight software-->
