+++ astrobee/launch/robot/LLP.launch
@@ -24,7 +24,7 @@
   <arg name="extra" default=""/>              <!-- Inject an additional node -->
   <arg name="debug" default=""/>              <!-- Debug a node set          -->
   <arg name="output" default="log"/>          <!-- Where to log              -->
-  <arg name="fake_localizer" default="false"/>    <!-- Runs fake_localizer -->
+  <arg name="ground_truth_localizer" default="false"/>    <!-- Runs ground_truth_localizer -->
 
   <!-- Launch all nodelet managers -->
   <group if="$(eval optenv('ASTROBEE_NODEGRAPH','')=='')">
@@ -45,7 +45,7 @@
   </group>
 
   <!-- Launch GNC nodes -->
-  <group unless="$(arg fake_localizer)">
+  <group unless="$(arg ground_truth_localizer)">
     <include file="$(find ff_util)/launch/ff_nodelet.launch">
       <arg name="class" value="imu_augmentor/ImuAugmentorNodelet" />
       <arg name="name" value="imu_augmentor_nodelet" />
