+++ localization/CMakeLists.txt
@@ -23,8 +23,8 @@ add_subdirectory(localization_measurements)
 add_subdirectory(sparse_mapping)
 
 if (USE_ROS)
-  add_subdirectory(fake_localizer)
   add_subdirectory(graph_localizer)
+  add_subdirectory(ground_truth_localizer)
   add_subdirectory(imu_augmentor)
   add_subdirectory(marker_tracking)
   add_subdirectory(lk_optical_flow)
