--- localization/fake_localizer/nodelet_plugins.xml
@@ -1,11 +1,0 @@
-<library path="lib/libfake_localizer">
-  <class name="fake_localizer/FakeLocalizerNodelet"
-	 type="fake_localizer::FakeLocalizerNodelet"
-	 base_class_type="nodelet::Nodelet">
-    <description>
-      This nodelet wraps fake localizer 
-    </description>
-  </class>
-</library>
-
-
