+++ localization/ground_truth_localizer/include/ground_truth_localizer/twist.h
@@ -15,23 +15,16 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-#ifndef FAKE_LOCALIZER_UTILITIES_H_
-#define FAKE_LOCALIZER_UTILITIES_H_
+#ifndef GROUND_TRUTH_LOCALIZER_TWIST_H_
+#define GROUND_TRUTH_LOCALIZER_TWIST_H_
 
-#include <fake_localizer/twist.h>
-#include <ff_msgs/EkfState.h>
-#include <localization_common/time.h>
+#include <Eigen/Core>
 
-#include <geometry_msgs/PoseStamped.h>
-#include <geometry_msgs/TwistStamped.h>
-#include <std_msgs/Header.h>
+namespace ground_truth_localizer {
+struct Twist {
+  Eigen::Vector3d linear_velocity;
+  Eigen::Vector3d angular_velocity;
+};
+}  // namespace ground_truth_localizer
 
-#include <Eigen/Geometry>
-
-namespace fake_localizer {
-Eigen::Isometry3d PoseFromMsg(const geometry_msgs::PoseStamped& pose_msg);
-Twist TwistFromMsg(const geometry_msgs::TwistStamped& twist_msg);
-ff_msgs::EkfState LocStateMsg(const Eigen::Isometry3d& pose, const Twist& twist, const localization_common::Time& time);
-}  // namespace fake_localizer
-
-#endif  // FAKE_LOCALIZER_UTILITIES_H_
+#endif  // GROUND_TRUTH_LOCALIZER_TWIST_H_
