+++ localization/ground_truth_localizer/readme.md
@@ -1,4 +1,4 @@
-\page fakelocalizer Fake Localizer
+\page groundtruthlocalizer Ground Truth Localizer
 
 # Package overview
 Echos ground truth poses and twists.
