+++ localization/ground_truth_localizer/src/ground_truth_localizer_nodelet.cc
@@ -16,44 +16,44 @@
  * under the License.
  */
 
-#include <fake_localizer/fake_localizer_nodelet.h>
-#include <fake_localizer/utilities.h>
+#include <ground_truth_localizer/ground_truth_localizer_nodelet.h>
+#include <ground_truth_localizer/utilities.h>
 #include <ff_msgs/EkfState.h>
 #include <ff_util/ff_names.h>
 #include <localization_common/utilities.h>
 
 #include <glog/logging.h>
 
-namespace fake_localizer {
+namespace ground_truth_localizer {
 namespace lc = localization_common;
-FakeLocalizerNodelet::FakeLocalizerNodelet()
+GroundTruthLocalizerNodelet::GroundTruthLocalizerNodelet()
     : ff_util::FreeFlyerNodelet(NODE_SIM_LOC, true), platform_name_(GetPlatform()) {}
 
-void FakeLocalizerNodelet::Initialize(ros::NodeHandle* nh) {
+void GroundTruthLocalizerNodelet::Initialize(ros::NodeHandle* nh) {
   ff_common::InitFreeFlyerApplication(getMyArgv());
   SubscribeAndAdvertise(nh);
 }
 
-void FakeLocalizerNodelet::SubscribeAndAdvertise(ros::NodeHandle* nh) {
+void GroundTruthLocalizerNodelet::SubscribeAndAdvertise(ros::NodeHandle* nh) {
   pose_pub_ = nh->advertise<geometry_msgs::PoseStamped>(TOPIC_LOCALIZATION_POSE, 1);
   twist_pub_ = nh->advertise<geometry_msgs::TwistStamped>(TOPIC_LOCALIZATION_TWIST, 1);
   state_pub_ = nh->advertise<ff_msgs::EkfState>(TOPIC_GNC_EKF, 1);
   heartbeat_pub_ = nh->advertise<ff_msgs::Heartbeat>(TOPIC_HEARTBEAT, 5, true);
 
-  pose_sub_ = nh->subscribe(TOPIC_LOCALIZATION_TRUTH, 1, &FakeLocalizerNodelet::PoseCallback, this,
+  pose_sub_ = nh->subscribe(TOPIC_LOCALIZATION_TRUTH, 1, &GroundTruthLocalizerNodelet::PoseCallback, this,
                             ros::TransportHints().tcpNoDelay());
-  twist_sub_ = nh->subscribe(TOPIC_LOCALIZATION_TRUTH_TWIST, 1, &FakeLocalizerNodelet::TwistCallback, this,
+  twist_sub_ = nh->subscribe(TOPIC_LOCALIZATION_TRUTH_TWIST, 1, &GroundTruthLocalizerNodelet::TwistCallback, this,
                              ros::TransportHints().tcpNoDelay());
 
-  input_mode_srv_ = nh->advertiseService(SERVICE_GNC_EKF_SET_INPUT, &FakeLocalizerNodelet::SetMode, this);
+  input_mode_srv_ = nh->advertiseService(SERVICE_GNC_EKF_SET_INPUT, &GroundTruthLocalizerNodelet::SetMode, this);
 }
 
-bool FakeLocalizerNodelet::SetMode(ff_msgs::SetEkfInput::Request& req, ff_msgs::SetEkfInput::Response& res) {
+bool GroundTruthLocalizerNodelet::SetMode(ff_msgs::SetEkfInput::Request& req, ff_msgs::SetEkfInput::Response& res) {
   input_mode_ = req.mode;
   return true;
 }
 
-void FakeLocalizerNodelet::PoseCallback(geometry_msgs::PoseStamped::ConstPtr const& pose) {
+void GroundTruthLocalizerNodelet::PoseCallback(geometry_msgs::PoseStamped::ConstPtr const& pose) {
   assert(pose->header.frame_id == "world");
   pose_ = PoseFromMsg(*pose);
   pose_pub_.publish(pose);
@@ -68,7 +68,7 @@ void FakeLocalizerNodelet::PoseCallback(geometry_msgs::PoseStamped::ConstPtr con
   heartbeat_pub_.publish(heartbeat_);
 }
 
-void FakeLocalizerNodelet::TwistCallback(geometry_msgs::TwistStamped::ConstPtr const& twist) {
+void GroundTruthLocalizerNodelet::TwistCallback(geometry_msgs::TwistStamped::ConstPtr const& twist) {
   assert(twist->header.frame_id == "world");
   twist_ = TwistFromMsg(*twist);
   twist_pub_.publish(twist);
@@ -76,7 +76,7 @@ void FakeLocalizerNodelet::TwistCallback(geometry_msgs::TwistStamped::ConstPtr c
   PublishLocState(timestamp);
 }
 
-void FakeLocalizerNodelet::PublishLocState(const lc::Time& timestamp) {
+void GroundTruthLocalizerNodelet::PublishLocState(const lc::Time& timestamp) {
   if (!twist_ || !pose_) return;
   const auto loc_state_msg = LocStateMsg(*pose_, *twist_, timestamp);
   state_pub_.publish(loc_state_msg);
@@ -85,6 +85,6 @@ void FakeLocalizerNodelet::PublishLocState(const lc::Time& timestamp) {
   const auto world_T_body_tf = lc::PoseToTF(*pose_, "world", "body", timestamp, platform_name_);
   transform_pub_.sendTransform(world_T_body_tf);
 }
-}  // namespace fake_localizer
+}  // namespace ground_truth_localizer
 
-PLUGINLIB_EXPORT_CLASS(fake_localizer::FakeLocalizerNodelet, nodelet::Nodelet);
+PLUGINLIB_EXPORT_CLASS(ground_truth_localizer::GroundTruthLocalizerNodelet, nodelet::Nodelet);
